changed float to double

+++ final/Foundation2/Order.cs
@@ -5,10 +5,10 @@ public class Order
     private List<string> products = new List<string>();
     private string _productName;
     private string _productID;
-    private float _productPrice;
+    private double _productPrice;
     private string _customerName;
     private string _customerAddress;
     private string _orderID;
 
-    
+
 }
